backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
eth-account==0.13.4
web3==7.6.0
supabase==2.10.0
httpx==0.27.2
redis==5.2.1
slowapi==0.1.9
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
